requirements.txt
filelock==3.18.0
fsspec==2025.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
pillow==11.2.1
sympy==1.14.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
